implemented gates

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -8,12 +8,21 @@ use App\Http\Traits\CanLoadRelationships;
 use App\Models\Attendee;
 use App\Models\Event;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class AttendeeController extends Controller
 {
     use CanLoadRelationships;
 
     private array $relations = ['user'];
+
+    public function __construct()
+    {
+        $this->middleware('auth:sanctum')->except(['index', 'show', 'update']);
+
+    }
+
+
     /**
      * Display a listing of the resource.
      */
@@ -65,8 +74,9 @@ class AttendeeController extends Controller
      */
     public function destroy(Event $event, Attendee $attendee)
     {
-        $attendee->delete();
+        Gate::authorize('delete-attendee', $event, $attendee);
 
+        $attendee->delete();
         return response(status: 204);
     }
 }
